requirements.txt
# Project Faraday - Python Dependencies
# Pinned major versions for PyInstaller stability

# Cryptography
argon2-cffi>=23,<25
cryptography>=41,<43

# Serialization
cbor2>=5.4,<6.0

# System tray (optional)
pystray>=0.19,<1.0
Pillow>=10.0,<11.0

